cliente/requirements.txt
Flask==3.1.0
requests==2.28.1
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.46
bcrypt==4.0.1
grpcio==1.59.3
grpcio-tools==1.59.3
pytest==7.1.3
PyJWT==2.4.0
cryptography==37.0.2
pybreaker==1.0.0